Add files via upload
Password_Cracker is based on Python Language. It uses two methods for cracking a Password . Common Password and Brute Force. For Common Password a txt file is uploaded here. The program reads the file and the input thus gives the number of attempts and Time in seconds was required. Where has Brute Force solely depends upon the processor speed of the device . As the name suggests it uses brute force method by guessing every character in the password and thus gives the number of attempts and Time in seconds was required. Its designed in simple GUI for user friendly ness.